522: v1.10: add criterion_pack, criterion_packs and update pack_list.html
